add context pacakge with WithResetableTimeout() functionality.
This commit adds a `context` package which contains an implementation
of a resetable-timeout context. This can be useful if you want to
extend the timeout of a given context which would be not trivially
possible otherwise.

To get an idea, look at this example:

```go
ctx, cancel := context.WithResetableTimeout(parent, 100*time.Millisecond)
defer cancel()
doWork() // for approx 80ms and decide then to take another 100ms to complete it
context.ResetTimeout(ctx, 100 * time.Millisecond)
doMoreWork() // which will not timeout in 20ms but in 100ms!
```