Improve err message of UUIDOrSQLName validator(#74)
Improve the error message of UUIDOrSQLName validator so that the message indicates that the value can be _either_ a valid UUID or a valid SQL name. The current message only shows the validation errors for the SQL name and this can confuse users because it implies that the UUID is not valid at all.