fix: do not parse body if ouput dest is nil (#135)
When an JSON api responses with 204 No Content, the BaseAPIClient
implementation would still attempt to parse the response body. But this
just throws and EOF error. When the API is responding with 204 this is
actually expected and the caller should/will pass a `nil` value to the
`out` parameter. The client implementation will now skip parsing the
body if the `out` destination is nil.

Signed-off-by: Lucas Roesler <roesler.lucas@gmail.com>