feat: Add request method to BaseAPIClient that expose the response object
Add the `Do` method to the `BaseAPIClient` to enable advanced use cases
that are not just the standard JSON REST API calls.

We will use this for a method that just wants to copy the response to
another Writer. Having direct access to the Response body provides the
least overhead.

Signed-off-by: Lucas Roesler <roesler.lucas@gmail.com>