feat: Add WithRetry to the BaseAPIClient interface (#155)
Add WithRetry to the BaseAPIClient interface. This allows the API client
to retry the request according the configured backoff plan and max attempts.

feat: Implement API task handler that uses the BaseAPIClient 

Create a task handler that replicates the behavior of api request
handler, but uses the BaseAPIClient interface instead of the stdlib
http.Client.

This allows us to implement the retry behavior via the new WithRetry
implementation of the BaseAPIClient.


fix: ensure taks process method stops the heartbeat goroutine

feat: add utility to simiplify testing for specific status codes

Add a utility function that helps users of the BaseAPIClient test for
specific status codes, by returning the http status code, if it can be
determined.