feat: implement heartbeat timeout in the worker (#167)
Allow a worker to abandon a task if the heartbeats are too slow, the current setting is one heartbeat every 15s. This will ensure that tasks handlers can not accidentally infinite loop and block the queue. Valid handlers _must_ heartbeat or the task will be failed. 

Add new Task Worker construction method to allow customization of this timeout period. 

Signed-off-by: Lucas Roesler <roesler.lucas@gmail.com>
Co-authored-by: Denis Rechkunov <mail@pragmader.me>