feat: allow customizing http handler error parsing (#270)
Expose the default Handler implementation and allow customizing the error parsing by passing a new parser function. This allows overriding the error handling but with a reasonable default.

Additionally, change ValidationErrorsToFiedErrorResponse signature to accept any map[string]error. This is the underlying type for a validation.Error, so it is backwards compatible, but it is also more flexible and will accept more type aliases for validation errors, making it easier to use.

Signed-off-by: Lucas Roesler <roesler.lucas@gmail.com>